+++ packages/adherev-util-resource/lib/resource.d.ts
@@ -1,2 +1,11 @@
-declare const _default;
+declare const _default: {
+    Dict: {
+        handlers: object;
+        value: any; /**
+         * initBasicDict
+         * Resource提供的字典命名规范 Resource${业务分组}${实际业务名称}
+         */
+        init: import("@baifendian/adhere-util-dict/lib/types").InitFunc;
+    };
+};
 export default _default;