+++ packages/adherev-ui-searchtable/changelog/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.0.4
+
+***
+
+2021-10-28
+
+* getFetchListPropNameToFirstUpper自动生成
+* column支持权限设置(authorized)
+* 加入renderTableHeader渲染(表格头上方，查询条件下方)
+* 加入renderTableFooter渲染(表格下方,分页上方)
+* searchtableimplement.tsx加入getSearchParams方法获取所有查询参数
+* 加入fixedTableSpaceBetween配置项(两端固定(表格的头始终在上方，分页始终在下方)
+* 修改无数据时样式的错乱
+* 修改loading时的样式错乱
+
 # 1.0.3
 
 ***